requirements.txt
amqp==2.3.2
billiard==3.5.0.3
celery==4.2.0
Django==1.11
django-extensions==2.0.7
djangorestframework==3.6.2
kombu==4.2.1
numpy==1.14.3
pymemcache==1.4.4
pytz==2018.4
scikit-learn==0.19.1
scipy==1.0.1
six==1.11.0
vine==1.1.4
